Unfinished Pathfinding 1.0.6

+++ Tiny Train Tales/Assets/Scripts/Manager/GameManager.cs
@@ -390,8 +390,12 @@ public class GameManager : MonoBehaviour
         PlayerPrefs.DeleteKey("Distance");
         PlayerPrefs.DeleteKey("RemainingDistance");
         PlayerPrefs.DeleteKey("Speed");
-        PlayerPrefs.DeleteKey("CurrentCity");
-        PlayerPrefs.DeleteKey("DestinationCity");
+
+        if (PlayerPrefs.HasKey("Finished"))
+        {
+            PlayerPrefs.DeleteKey("CurrentCity");
+            PlayerPrefs.DeleteKey("DestinationCity");
+        }
     }
 
     public void SaveAll()
